Add some simple tests (and a structure to define more)
Currently it only has a very simple test but hopefully in the future
it will expand to cover everything more thoroughly. Also, it needs
some form of automated checking if the test worked; and possibly a
better format for defining more tests (this might be easiest with a
script).

From now on, a new test should be made for every bug.